+++ src/term_image/exceptions.py
@@ -6,6 +6,10 @@ Custom Exceptions
 from __future__ import annotations
 
 
+class TermImageWarning(UserWarning):
+    """Package-specific warning category."""
+
+
 class TermImageError(Exception):
     """Exception baseclass. Raised for generic errors."""
 
@@ -66,7 +70,7 @@ class KittyImageError(GraphicsImageError):
     """
 
 
-__all__ = [
+__all__ = ["TermImageWarning"] + [
     name
     for name, obj in vars().items()
     if isinstance(obj, type) and issubclass(obj, TermImageError)